requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
python-whois>=0.8.0
geoip2>=4.7.0
networkx>=3.2.0
matplotlib>=3.8.0
pyvis>=0.3.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.1
rich>=13.7.0
jinja2>=3.1.2
ipwhois>=1.2.0
dnspython>=2.4.0
tqdm>=4.66.0
